Add part of recirc load back to primary in load shift
Alters the primary heating capactiy assuming the primary component has to cover 50 W/apt and converting that to extra gallons of capactiy needed. Which also goes back to affecting the running volume of the storage tank. The running vollume is calculated assuming a worst case scenario where the primary storage is removed as if it was at the 120 F

- Tests updated